fixed user controller
