+++ sk0-back/src/users/controllers/users.controller.ts
@@ -1,0 +1,27 @@
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  UsePipes,
+  ValidationPipe,
+} from '@nestjs/common';
+import { CreateUserDto } from '../dtos/create-user.dto';
+import { UsersService } from '../services/users.service';
+
+@Controller('users')
+export class UsersController {
+  constructor(private usersService: UsersService) {}
+
+  @Get('')
+  getUsers() {
+    return this.usersService.getUsers();
+  }
+
+  @Post('register')
+  @UsePipes(ValidationPipe)
+  async createUser(@Body() createUserDto: CreateUserDto) {
+    await this.usersService.createUser(createUserDto);
+    return { message: 'User created successfully' };
+  }
+}
