+++ sk0-back/src/users/services/users.service.ts
@@ -1,13 +1,12 @@
 import { Injectable } from '@nestjs/common';
-import { CreateUserDto } from 'src/users/dtos/createUser.dto';
-import { User } from 'src/users/types/User';
+import { CreateUserDto } from '../dtos/create-user.dto';
+import { User } from '../types/user';
 
 @Injectable()
 export class UsersService {
   private users: User[] = [
     {
-      id: 2,
-      username: 'testuser',
+      name: 'testuser',
       email: 'testuser@example.com',
       password: 'password1234',
     },
@@ -21,7 +20,7 @@ export class UsersService {
     return this.users;
   }
 
-  getByEmail(email: string): User | undefined {
+  getByEmail(email: string): User | null {
     return this.users.find((user) => user.email === email);
   }
 }
