+++ sk0-back/src/users/types/user.ts
@@ -1,6 +1,5 @@
 export interface User {
-  id: number;
-  username: string;
+  name: string;
   email: string;
   password: string;
 }
